fix: ipc with electron process
