+++ packages/main-process/src/parts/AppWindow/AppWindow.js
@@ -8,7 +8,6 @@ const Platform = require('../Platform/Platform.js')
 const Preferences = require('../Preferences/Preferences.js')
 const AppWindowStates = require('../AppWindowStates/AppWindowStates.js')
 const Logger = require('../Logger/Logger.js')
-const Electron = require('electron')
 const ElectronApplicationMenu = require('../ElectronApplicationMenu/ElectronApplicationMenu.js')
 
 // TODO impossible to test these methods
