+++ packages/main-process/src/parts/ContentSecurityPolicy/ContentSecurityPolicy.js
@@ -13,6 +13,7 @@ exports.value = [
   `script-src 'self'`,
   `media-src 'self'`,
   `prefetch-src 'self'`,
+  `manifest-src 'self'`,
   `style-src 'self' 'unsafe-inline'`, // TODO remove unsafe-inline
 ]
   .map(addSemicolon)
