+++ packages/renderer-process/src/parts/IpcParent/IpcParent.js
@@ -10,6 +10,8 @@ const getModule = (method) => {
       return import('./IpcParentWithReferencePort.js')
     case IpcParentType.ModuleWorkerWithMessagePort:
       return import('./IpcParentWithModuleWorkerWithMessagePort.js')
+    case IpcParentType.Electron:
+      return import('./IpcParentWithElectron.js')
     default:
       throw new Error('unexpected ipc type')
   }
